refactor: move types to the component files
