+++ src/components/messageSpace/messageSpace.tsx
@@ -4,8 +4,16 @@ import Box from '@mui/system/Box'
 import React from 'react'
 
 import ElementRenderer from '../elementRenderer/elementRenderer'
-import MessageCanvas from '../messageCanvas/messageCanvas'
-import type { MessageSpaceProps } from '../types'
+import MessageCanvas, {
+  type MessageContainerProps,
+} from '../messageCanvas/messageCanvas'
+import type { ComponentMap, ThreadableMessage } from '../types'
+
+export interface MessageSpaceProps extends MessageContainerProps {
+  /** A component map contains message formats as keys and their corresponding React components as values. */
+  supportedElements: ComponentMap
+  messages?: ThreadableMessage[]
+}
 
 /**
  The `MessageSpace` component uses `MessageCanvas` and `ElementRenderer` to render a list of messages. It serves as a container for individual message items, each encapsulated within a `MessageCanvas` for consistent styling and layout. \n\n Note: For more information about the `getActionsComponent` and `getProfileComponent` fields, refer to the [MessageCanvas' docs](http://localhost:6006/?path=/docs/rustic-ui-message-canvas-message-canvas--docs).
